rename: myMission -> submissions로 변경 및 파일, 네이밍 변경
