+++ frontend/src/apis/paths.ts
@@ -1,5 +1,5 @@
 export const PATH = {
   missionList: '/missions',
   submissions: '/submissions',
-  missionInProgress: '/submissions/now',
+  submissionsInProgress: '/submissions/now',
 };
