+++ frontend/src/components/Submission/SubmissionCompleted.tsx
@@ -1,0 +1,11 @@
+import SubmissionCompletedCard from './SubmissionCompletedCard';
+import * as S from './Submission.styled';
+
+export default function SubmissionCompleted() {
+  return (
+    <S.SubmissionCompletedContainer>
+      <S.CompletedTitle>완료한 미션</S.CompletedTitle>
+      <SubmissionCompletedCard />
+    </S.SubmissionCompletedContainer>
+  );
+}
