+++ frontend/src/hooks/useMissionCompleted.ts
@@ -1,0 +1,11 @@
+// import { useSuspenseQuery } from "@tanstack/react-query"
+// import { missionKeys } from "./queries/keys"
+
+// const useMissionCompleted = () => {
+//   return useSuspenseQuery<>({
+//     queryKey: missionKeys.completed,
+//     queryFn: 
+//   })
+// }
+
+// export default useMissionCompleted
