+++ frontend/src/hooks/useMissionInProgress.ts
@@ -1,10 +1,10 @@
 import { getMissionInProgress } from '@/apis/missionAPI';
-import type { MissionInProgress } from '@/types';
+import type { MissionSubmission } from '@/types';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { missionKeys } from './queries/keys';
 
 const useMissionInProgress = () => {
-  return useSuspenseQuery<MissionInProgress>({
+  return useSuspenseQuery<MissionSubmission>({
     queryKey: missionKeys.inProgress,
     queryFn: getMissionInProgress,
   });
