+++ frontend/src/index.tsx
@@ -7,7 +7,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import MissionDetailPage from './pages/MissionDetailPage';
 import MissionListPage from './pages/MissionListPage';
 import MissionSubmitPage from './pages/MissionSubmitPage';
-import MyMissionPage from './pages/MyMissionPage';
+import SubmissionPage from './pages/SubmissionPage';
 import UserProfilePage from './pages/UserProfilePage';
 import GuidePage from './pages/GuidePage';
 import React, { Suspense } from 'react';
@@ -71,10 +71,10 @@ const routes = [
     ),
   },
   {
-    path: ROUTES.myMission,
+    path: ROUTES.submissions,
     element: (
       <App>
-        <MyMissionPage />
+        <SubmissionPage />
       </App>
     ),
   },
