+++ frontend/src/types/index.ts
@@ -27,7 +27,7 @@ export interface Mission {
   url: string;
 }
 
-export interface MissionInProgress {
+export interface MissionSubmission {
   id: number;
   mission: Mission;
   myPrLink: string;
